sua admin

+++ js/cart_admin.js
@@ -14,7 +14,7 @@ function cart_admin() {
                         </select>
                     </div>
                     <div class="header_seach">
-                        <input type="text" name="" id="" placeholder="Tìm Kiếm Đơn Hàng"/>
+                        <input type="text" name="" id="" placeholder="Tìm Kiếm Mã Đơn Hàng"/>
                     </div>
                     <div class="date_1">
                         <label for="startDate">Từ</label>
@@ -26,7 +26,6 @@ function cart_admin() {
                     </div>
                     <button onclick="searchByDate()">Tìm kiếm</button>
                     <div class="reset">
-                        <button>Reset</button>
                     </div>
                 </div>
                 <div class="table_cart">
@@ -76,6 +75,7 @@ function cart_admin() {
         right_header.innerHTML = cart_list;
         oder_user();
         pregnancy_status();
+        idproduct();
     });
 }
 cart_admin();
@@ -116,7 +116,6 @@ function order_view(id) {
     view();
     const orders = JSON.parse(localStorage.getItem('orders'));
     let cart_item = document.getElementsByClassName('cart_item')[0];
-    var id_user = orders[id].Trangthai;
     console.log(orders[id].Sanpham);
 
     var list_product_cart = '';
@@ -169,7 +168,7 @@ function order_view(id) {
                     <h4>Thời Gian Đặt :</h4>
                 </div>
             </td>
-            <td>${orders[id].user.date}</td>
+            <td>${orders[id].Ngaydat}</td>
         </tr>
         <tr>
             <td>
@@ -190,18 +189,30 @@ function order_view(id) {
                 </div>
             </td>
             <td style="color:red">${giaTienChinhThuc}</td>
+        
         </tr>
-        <button id="btn_handle">${orders[id].Trangthai}</button>
+       <tr>
+       <button id="btn_handle">${orders[id].Trangthai}</button>
+       <button onclick="reset(${id})">ok</button>
+       </tr>
+        
         `;
     table_threed.innerHTML = user_table;
     document.getElementById('btn_handle').addEventListener('click', () => {
-        var btn_handle = document.getElementById('btn_handle');
-        btn_handle.textContent = 'Đã xử lý';
-        orders[id].Trangthai = btn_handle.textContent;
-        btn_handle.style.background = '#0083d6';
-        btn_handle.style.color = 'white';
-        localStorage.setItem('orders', JSON.stringify(orders));
-        oder_user();
+        if (orders[id].Trangthai === 'Đã xử lý') {
+            return;
+        }
+
+        let handle = confirm('Bạn Có Chắc Xử Lý Không ?');
+        if (handle) {
+            var btn_handle = document.getElementById('btn_handle');
+            btn_handle.textContent = 'Đã xử lý';
+            orders[id].Trangthai = btn_handle.textContent;
+            btn_handle.style.background = '#0083d6';
+            btn_handle.style.color = 'white';
+            localStorage.setItem('orders', JSON.stringify(orders));
+            oder_user();
+        }
     });
     if (orders[id].Trangthai === 'Đã xử lý') {
         btn_handle.style.background = '#0083d6';
@@ -212,7 +223,6 @@ function order_view(id) {
 //selec product
 function pregnancy_status() {
     let select_user = document.getElementById('select_userall');
-
     select_user.addEventListener('change', () => {
         if (select_user.value === 'Tất cả') {
             oder_user();
@@ -227,7 +237,6 @@ function oder_userByStatus(status) {
     const orders = JSON.parse(localStorage.getItem('orders'));
     let tbody = document.getElementsByTagName('tbody')[0];
     let table_user = '';
-
     orders.forEach((list, index) => {
         if (list.Trangthai === status) {
             let row = `
@@ -237,7 +246,10 @@ function oder_userByStatus(status) {
                     <td>${list.Ngaydat}</td>
                     <td>${list.Tongtien}</td>
                     <td><button class="btn_user">${list.Trangthai}</button></td>
-                    <td><button class="btn_view" onclick="order_view(${index})">Xem</button></td>
+                    <td><button class="btn_view" onclick="order_view(${index})">Xem</button>
+                   
+                    </td>
+                    
                 </tr>
             `;
             table_user += row;
@@ -255,16 +267,85 @@ function oder_userByStatus(status) {
 function searchByDate() {
     const startDate = new Date(document.getElementById('startDate').value);
     const endDate = new Date(document.getElementById('endDate').value);
+    const orders = JSON.parse(localStorage.getItem('orders'));
+    let tbody = document.getElementsByTagName('tbody')[0];
 
-    var datesInRange = [];
-    for (var currentDate = new Date(startDate); currentDate <= endDate; currentDate.setDate(currentDate.getDate() + 1)) {
-        var date = new Date(currentDate);
-        var day = date.getDate();
-        var month = date.getMonth() + 1;
-        var year = date.getFullYear();
-        var dateFormatted = month + '/' + day + '/' + year;
-        datesInRange.push(dateFormatted);
-    }
+    var a = endDate.getDate();
+    var b = endDate.getFullYear();
+    var c = endDate.getMonth() + 1;
+    let sumenDate = a + b + c;
+    var d = startDate.getDate();
+    var e = startDate.getFullYear();
+    var f = startDate.getMonth() + 1;
+    let sumstartDate = d + e + f;
+    var tong = 0;
+    let table_user = '';
+    orders.map((list, index) => {
+        var listdate = list.Ngaydat.split('/');
+        const day = parseInt(listdate[0], 10);
+        const month = parseInt(listdate[1], 10);
+        const year = parseInt(listdate[2], 10);
+        const sum = day + month + year;
+        if (sumstartDate <= sum && sumenDate >= sum) {
+            let row = `
+            <tr>
+                <td>${index + 1}</td>
+                <td>${list.user.name}</td>
+                <td>${list.Ngaydat}</td>
+                <td>${list.Tongtien}</td>
+                <td><button class="btn_user">${list.Trangthai}</button></td>
+                <td><button class="btn_view" onclick="order_view(${index})">Xem</button>
+               
+                </td>
+            </tr>
+        `;
+            table_user += row;
+        }
+    });
+    tbody.innerHTML = table_user;
+}
+
+// Tìm Kiếm Mã Đơn Hàng
+function idproduct() {
+    const orders = JSON.parse(localStorage.getItem('orders'));
+    let header_seach = document.querySelectorAll('.header_seach input')[0];
+    let tbody = document.getElementsByTagName('tbody')[0];
+    header_seach.addEventListener('keydown', (event) => {
+        if (event.key === 'Enter') {
+            let table_user = '';
+            let found = false;
+            orders.forEach((list, index) => {
+                let orderid = list.madon.toUpperCase();
+                let value_input = header_seach.value.toUpperCase();
+                if (orderid === value_input) {
+                    found = true;
+                    let row = `
+                        <tr>
+                            <td>${list.madon}</td>
+                            <td>${list.user.name}</td>
+                            <td>${list.Ngaydat}</td>
+                            <td>${list.Tongtien}</td>
+                            <td><button class="btn_user">${list.Trangthai}</button></td>
+                            <td><button class="btn_view" onclick="order_view(${index})">Xem</button></td>
+                        </tr>
+                    `;
+                    table_user += row;
+                }
+            });
+            if (!found) {
+                table_user = `<tr><td colspan="6">Không Tìm Thấy Đơn Hàng Nào Cả</td></tr>`;
+            }
+
+            tbody.innerHTML = table_user;
+            let btn_users = document.getElementsByClassName('btn_user');
+            Array.from(btn_users).forEach((btn_user) => {
+                if (btn_user.textContent === 'Đã xử lý') {
+                    btn_user.style.background = '#0083d6';
+                    btn_user.style.color = 'white';
+                }
+            });
+        }
+    });
 }
 
 function closs_order() {
@@ -284,7 +365,6 @@ function view() {
     cart_order.classList.add('top');
     setOpacityAndDisableEvents();
 }
-
 function setOpacityAndDisableEvents() {
     const elements = document.querySelectorAll(
         '.table_cart  , .logo, .left_list, .list_product, .selection_food, .pagination-buttons,.header_cart_table',
@@ -293,4 +373,61 @@ function setOpacityAndDisableEvents() {
         element.style.opacity = '0.5';
         element.style.pointerEvents = 'none';
     });
+}
+
+///file word
+function reset(id) {
+    const orders = JSON.parse(localStorage.getItem('orders'));
+    if (orders && orders[id]) {
+        const { madon, tenNguoiNhan, soDienThoai, diaChi, tenSanPham, gia, ngayDat } = orders[id];
+
+        const doc = new window.docx.Document();
+
+        const titleStyle = {
+            paragraph: {
+                alignment: window.docx.AlignmentType.CENTER,
+                spacing: {
+                    before: 200,
+                    after: 200,
+                },
+            },
+            run: {
+                bold: true,
+                size: 28,
+                color: '2F5496',
+            },
+        };
+
+        const contentStyle = {
+            paragraph: {
+                spacing: {
+                    before: 100,
+                    after: 200,
+                },
+            },
+            run: {
+                size: 24,
+                color: '000000',
+            },
+        };
+
+        doc.addSection({
+            children: [
+                new window.docx.Paragraph('Chi tiết Đơn Hàng', titleStyle),
+                new window.docx.Paragraph('Mã Đơn: ' + madon, contentStyle),
+                new window.docx.Paragraph('Tên Người Nhận: ' + tenNguoiNhan, contentStyle),
+                new window.docx.Paragraph('Số Điện Thoại: ' + soDienThoai, contentStyle),
+                new window.docx.Paragraph('Địa Chỉ: ' + diaChi, contentStyle),
+                new window.docx.Paragraph('Đơn Hàng: ' + tenSanPham, contentStyle),
+                new window.docx.Paragraph('Giá: ' + gia + ' vnd', contentStyle),
+                new window.docx.Paragraph('Ngày Đặt: ' + ngayDat, contentStyle),
+            ],
+        });
+
+        window.docx.Packer.toBlob(doc).then((blob) => {
+            window.saveAs(blob, 'chi_tiet_don_hang.docx');
+        });
+    } else {
+        console.error(`Không tìm thấy dữ liệu cho ID: ${id}`);
+    }
 }
